+++ formula-android/src/main/java/com/instacart/formula/android/ActivityConfigurator.kt
@@ -1,8 +1,6 @@
 package com.instacart.formula.android
 
 import androidx.fragment.app.FragmentActivity
-import com.instacart.formula.integration.ActivityStore
-import com.instacart.formula.integration.ActivityStoreContext
 import kotlin.reflect.KClass
 
 /**