+++ formula-android/src/main/java/com/instacart/formula/android/FormulaAppCompatActivity.kt
@@ -1,4 +1,4 @@
-package com.instacart.formula.integration
+package com.instacart.formula.android
 
 import android.content.Intent
 import android.os.Bundle