+++ formula-android/src/main/java/com/instacart/formula/android/internal/ActivityManager.kt
@@ -2,10 +2,9 @@ package com.instacart.formula.android.internal
 
 import androidx.fragment.app.FragmentActivity
 import androidx.lifecycle.Lifecycle
-import com.instacart.formula.activity.ActivityResult
+import com.instacart.formula.android.events.ActivityResult
 import com.instacart.formula.fragment.FragmentEnvironment
-import com.instacart.formula.integration.ActivityStore
-import com.instacart.formula.integration.FragmentFlowRenderView
+import com.instacart.formula.android.ActivityStore
 import io.reactivex.rxjava3.disposables.CompositeDisposable
 import io.reactivex.rxjava3.disposables.Disposable