+++ formula-android/src/main/java/com/instacart/formula/integration/FeatureProvider.kt
@@ -1,5 +1,0 @@
-package com.instacart.formula.integration
-
-internal interface FeatureProvider {
-    fun getFeature(id: FragmentId): FeatureEvent?
-}