+++ CHANGELOG.md
@@ -17,6 +17,12 @@
 - **Breaking**: Replacing `FlowDeclaration` and `FlowIntegration` with `FlowFactory`.
 - Added `bind(myFeatureFactory) { component -> component.myFeatureDependency() }` binding method.
 - Added `bind(myFlowFactory) { component -> component.createMyFlowDependencies() }` binding method.
+- Rename `KeyState` to `FragmentState` and move it to `android` package.
+- Move `ActivityStoreContext` into `android` package.
+- Move `ActivityStore` into `android` package.
+- Move `FragmentId` into `android` package.
+- Move `FormulaAppCompatActivity` into `android` package.
+- Move `ActivityResult` into `android.events` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.